+++ src/communities/searchSummary.ts
@@ -20,7 +20,7 @@ import queryBuilder from "./queryBuilder"
  * @param fetchExecutor An instance of the request executor.
  * @returns An array of mobile home results.
  */
-async function summary(
+async function searchSummary(
   params: ISearchParams,
   creds: ICredentials,
   environment?: IEnvironment,
@@ -56,4 +56,4 @@ async function summary(
     ) || []
   )
 }
-export default summary
+export default searchSummary
